Move filter function back to DataFrame and add it to Series.
