+++ databricks/koalas/generic.py
@@ -17,7 +17,6 @@
 """
 A base class to be monkey-patched to DataFrame/Column to behave similar to pandas DataFrame/Series.
 """
-import re
 import warnings
 from collections import Counter
 from collections.abc import Iterable
@@ -27,7 +26,6 @@ from typing import Optional, Union, List
 
 import numpy as np
 import pandas as pd
-from pandas.api.types import is_list_like
 
 from pyspark import sql as spark
 from pyspark.sql import functions as F
@@ -1635,151 +1633,6 @@ class Frame(object):
             first_valid_idx = first_valid_idx[0]
 
         return first_valid_idx
-
-    def filter(self, items=None, like=None, regex=None, axis=None):
-        """
-        Subset rows or columns of dataframe according to labels in
-        the specified index.
-        Note that this routine does not filter a dataframe on its
-        contents. The filter is applied to the labels of the index.
-        Parameters
-        ----------
-        items : list-like
-            Keep labels from axis which are in items.
-        like : string
-            Keep labels from axis for which "like in label == True".
-        regex : string (regular expression)
-            Keep labels from axis for which re.search(regex, label) == True.
-        axis : int or string axis name
-            The axis to filter on.  By default this is the info axis,
-            'index' for Series, 'columns' for DataFrame.
-        Returns
-        -------
-        same type as input object
-        See Also
-        --------
-        DataFrame.loc
-        Notes
-        -----
-        The ``items``, ``like``, and ``regex`` parameters are
-        enforced to be mutually exclusive.
-        ``axis`` defaults to the info axis that is used when indexing
-        with ``[]``.
-        Examples
-        --------
-        >>> df = ks.DataFrame(np.array(([1, 2, 3], [4, 5, 6])),
-        ...                   index=['mouse', 'rabbit'],
-        ...                   columns=['one', 'two', 'three'])
-        >>> # select columns by name
-        >>> df.filter(items=['one', 'three'])
-                one  three
-        mouse     1      3
-        rabbit    4      6
-        >>> # select columns by regular expression
-        >>> df.filter(regex='e$', axis=1)
-                one  three
-        mouse     1      3
-        rabbit    4      6
-        >>> # select rows containing 'bbi'
-        >>> df.filter(like='bbi', axis=0)
-                one  two  three
-        rabbit    4    5      6
-        """
-        from databricks.koalas.series import first_series
-
-        if sum(x is not None for x in (items, like, regex)) > 1:
-            raise TypeError(
-                "Keyword arguments `items`, `like`, or `regex` " "are mutually exclusive"
-            )
-
-        is_series = isinstance(self, ks.Series)
-
-        kdf = self
-        if is_series:
-            kdf = self.to_frame()
-            axis = validate_axis(axis)
-            if axis == 1:
-                raise ValueError("Series does not support columns axis.")
-        else:
-            axis = validate_axis(axis, none_axis=1)
-        assert isinstance(kdf, ks.DataFrame)
-
-        index_scols = kdf._internal.index_spark_columns
-
-        if items is not None:
-            if is_list_like(items):
-                items = list(items)
-            else:
-                raise ValueError("items should be a list-like object.")
-            if axis == 0:
-                if len(index_scols) == 1:
-                    col = None
-                    for item in items:
-                        if col is None:
-                            col = index_scols[0] == F.lit(item)
-                        else:
-                            col = col | (index_scols[0] == F.lit(item))
-                elif len(index_scols) > 1:
-                    # for multi-index
-                    col = None
-                    for item in items:
-                        if not isinstance(item, (tuple)):
-                            raise TypeError("Unsupported type {}".format(type(item)))
-                        if not item:
-                            raise ValueError("The item should not be empty.")
-                        midx_col = None
-                        for i, element in enumerate(item):
-                            if midx_col is None:
-                                midx_col = index_scols[i] == F.lit(element)
-                            else:
-                                midx_col = midx_col & (index_scols[i] == F.lit(element))
-                        if col is None:
-                            col = midx_col
-                        else:
-                            col = col | midx_col
-                else:
-                    raise ValueError("Single or multi index must be specified.")
-                filtered_df = ks.DataFrame(kdf._internal.with_filter(col))
-            elif axis == 1:
-                filtered_df = kdf[items]
-        elif like is not None:
-            if axis == 0:
-                col = None
-                for index_scol in index_scols:
-                    if col is None:
-                        col = index_scol.contains(like)
-                    else:
-                        col = col | index_scol.contains(like)
-                filtered_df = ks.DataFrame(self._internal.with_filter(col))
-            elif axis == 1:
-                column_labels = kdf._internal.column_labels
-                output_labels = [label for label in column_labels if any(like in i for i in label)]
-                filtered_df = kdf[output_labels]
-        elif regex is not None:
-            if axis == 0:
-                col = None
-                for index_scol in index_scols:
-                    if col is None:
-                        col = index_scol.rlike(regex)
-                    else:
-                        col = col | index_scol.rlike(regex)
-                filtered_df = ks.DataFrame(self._internal.with_filter(col))
-            elif axis == 1:
-                column_labels = kdf._internal.column_labels
-                matcher = re.compile(regex)
-                output_labels = [
-                    label
-                    for label in column_labels
-                    if any(matcher.search(i) is not None for i in label)
-                ]
-                filtered_df = kdf[output_labels]
-        else:
-            raise TypeError("Must pass either `items`, `like`, or `regex`")
-
-        if is_series:
-            return first_series(filtered_df)
-        else:
-            return filtered_df
 
     def median(self, axis=None, numeric_only=True, accuracy=10000):
         """
